+++ openpls/unidimensionality.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python3
 #
 # Copyright (C) 2019 Google Inc.
+# Copyright (C) 2026 Johannes Jacob / OpenPLS
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -15,66 +16,33 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import numpy as np
-import pandas as pd
-from sklearn.decomposition import PCA
+"""Backwards-compatibility shim for the pre-1.9.0 ``Unidimensionality`` name.
 
-import openpls.util as util
-from openpls.config import Config
-from openpls.mode import Mode
+The class was renamed to :class:`openpls.reliability.Reliability` in 1.9.0
+because Cronbach's alpha and Dillon-Goldstein rho_c measure reliability,
+not unidimensionality. Import the new name; ``Unidimensionality`` will be
+removed in a future release.
+"""
+
+import warnings
+
+from openpls.reliability import Reliability
 
 
-class Unidimensionality:
-    """Internal class that computes various reliability metrics. Use the method :meth:`~.openpls.Plspm.unidimensionality` defined on :class:`~.openpls.Plspm` to retrieve the results."""
-    def __init__(self, config: Config, data: pd.DataFrame, correction: float):
-        self.__config = config
-        self.__data = data
-        self.__correction = correction
+class Unidimensionality(Reliability):
+    """Deprecated alias for :class:`openpls.reliability.Reliability`.
 
-    def summary(self):
-        """Internal method that performs principal component analysis to compute various reliability metrics.
+    Kept so external callers importing ``openpls.unidimensionality.Unidimensionality``
+    continue to work; emits a :class:`DeprecationWarning` on construction.
+    """
 
-        For latent variables whose indicator block contains missing values, the
-        computation falls back to listwise deletion (rows with any NaN in the
-        block are dropped) with a per-block correction factor. Upstream `plspm`
-        skipped the LV entirely in this case (OpenPLS fix #65).
-        """
-        summary = pd.DataFrame({"mode":                 pd.Series(dtype="str"),
-                                "mvs":                  pd.Series(dtype="float"),
-                                "cronbach_alpha":       pd.Series(dtype="float"),
-                                "dillon_goldstein_rho": pd.Series(dtype="float"),
-                                "eig_1st":              pd.Series(dtype="float"),
-                                "eig_2nd":              pd.Series(dtype="float")},
-                                 index=list(self.__config.path()))
-        for lv in list(self.__config.path()):
-            mvs = len(self.__config.mvs(lv))
-            summary.loc[lv, "mode"] = self.__config.mode(lv).name
-            summary.loc[lv, "mvs"] = mvs
-            block = self.__data.loc[:, self.__config.mvs(lv)]
-            if block.isnull().values.any():
-                block = block.dropna()
-                if block.shape[0] < 2:
-                    continue
-                correction = np.sqrt(block.shape[0] / (block.shape[0] - 1))
-            else:
-                correction = self.__correction
-            mvs_for_lvs = util.treat(block) * correction
-            pca_input = mvs_for_lvs if mvs_for_lvs.shape[0] > mvs_for_lvs.shape[1] else mvs_for_lvs.transpose()
-            pca = PCA()
-            pca_scores = pca.fit_transform(pca_input)
-            pca_std_dev = np.std(pca_scores, axis=0)
-            summary.loc[lv, "eig_1st"] = pca_std_dev[0] ** 2
-            summary.loc[lv, "eig_2nd"] = pca_std_dev[1] ** 2 if mvs > 1 else np.nan
-            if (self.__config.mode(lv) == Mode.A):
-                if mvs > 1:
-                    ca_numerator = 2 * np.tril(pca_input.corr(), -1).sum()
-                    ca_denominator = pca_input.sum(axis=1).var() / correction ** 2
-                    ca = max(0, (ca_numerator / ca_denominator) * (mvs / (mvs - 1)))
-                else:
-                    ca = np.nan
-                summary.loc[lv, "cronbach_alpha"] = ca
-                corr = np.corrcoef(np.column_stack((pca_input.values, pca_scores[:,0])), rowvar=False)[:,-1][:-1]
-                rho_numerator = sum(corr) ** 2
-                rho_denominator = rho_numerator + (mvs - np.sum(np.power(corr, 2)))
-                summary.loc[lv, "dillon_goldstein_rho"] = rho_numerator / rho_denominator
-        return summary
+    def __init__(self, *args, **kwargs):
+        warnings.warn(
+            "openpls.unidimensionality.Unidimensionality is deprecated; "
+            "use openpls.reliability.Reliability instead. The metrics it "
+            "computes (Cronbach's alpha, Dillon-Goldstein rho_c) measure "
+            "reliability, not unidimensionality.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        super().__init__(*args, **kwargs)